Produce a more PEP8-compliant project
